requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
django-jazzmin==2.4.9
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
sweetify==2.3.0
